Fix #173: Reorganize bindings
 - Move the bindings to the org.scala-native.bindings organization.
 - Move binding modules under org.scalanative.bindings dropping the
   per-binding package if only one module is emitted.
 - Drop the "lib" prefix from the binding artifact name.
