+++ bindings/iconv/src/main/scala/org/scalanative/bindings/iconv.scala
@@ -1,4 +1,4 @@
-package org.scalanative.bindgen.bindings.iconv
+package org.scalanative.bindings
 
 import scala.scalanative._
 import scala.scalanative.native._
