+++ bindings/posix/src/test/scala/org/scalanative/bindings/tests/FnmatchSpec.scala
@@ -1,4 +1,4 @@
-package org.scalanative.bindgen.bindings.tests
+package org.scalanative.bindings.tests
 
 import org.scalatest.FunSpec
 
@@ -7,7 +7,7 @@ class FnmatchSpec extends FunSpec {
     it("should match patterns") {
       //#usage-example
       import scala.scalanative.native._
-      import org.scalanative.bindgen.bindings.posix.fnmatch._
+      import org.scalanative.bindings.posix.fnmatch._
 
       assert(fnmatch(c"*.md", c"README.md", 0) == 0)
       assert(fnmatch(c"*.[ch]", c"main.h", 0) == 0)
