+++ bindings/utf8proc/src/main/scala/org/scalanative/bindings/utf8proc.scala
@@ -1,4 +1,4 @@
-package org.scalanative.bindgen.bindings.utf8proc
+package org.scalanative.bindings
 
 import scala.scalanative._
 import scala.scalanative.native._
